apps/server/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.0.2
packaging==24.2
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rank-bm25==0.2.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
slugify==0.0.1
soupsieve==2.8
SQLAlchemy==2.0.43
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
